added single route for edit and post session
