+++ src/pages/api/PostFormData.ts
@@ -1,3 +1,4 @@
+import { PatchTypes } from "@/types/ResponseTypes";
 import { RowType } from "@/types/types";
 import { publishMessage } from "@/utils/PublishSnsMessage";
 import { instance } from "@/utils/RootClient";
@@ -25,21 +26,20 @@ async function postFormDataToBackend(formData: RowType) {
   );
 
   const requestBody = {
-    name: test.name, //x
-    platform: test.platform, //x
-    platform_link: "", //x
-    portal_link: "", //responds to the portal link //x
+    name: test.name,
+    platform: test.platform,
+    platform_link: "",
+    portal_link: "",
     start_time: start_time,
     end_time: end_time,
     is_active: "",
     purpose: {
-      //x
       type: "attendance",
       params: "quiz",
     },
     repeat_schedule: { type: "weekly", params: [1, 2, 3, 4, 5, 6, 7] }, // this is hardcoded needs to be changed
-    session_id: "", //x
-    platform_id: "", //x
+    session_id: "",
+    platform_id: "",
     // form_schema_id: session.form_schema_id, // should be signup_form_name
     signup_form_name: "Haryana Registration Form",
     type: "sign-in",
@@ -50,14 +50,14 @@ async function postFormDataToBackend(formData: RowType) {
     pop_up_form: session.pop_up_form,
     number_of_fields_in_pop_form: session.number_of_fields_in_pop_form,
     meta_data: {
-      group: student.program, //x
+      group: student.program,
       batch: student.batch,
       grade: student.grade,
       course: student.course,
       stream: student.stream,
       test_format: test.format,
       test_purpose: test.purpose,
-      enabled: timeline.isEnabled, //remove it as it not used
+      enabled: timeline.isEnabled,
       cms_test_id:
         // "https://cms.peerlearning.com/chapter_tests/655df9a23562d97a6300b53e",
         test.cmsId, //x
@@ -66,10 +66,10 @@ async function postFormDataToBackend(formData: RowType) {
       optional_limits: test.optionalLimit,
       marking_scheme: test.markingScheme,
       test_type: test.type,
-      shortened_link: "", //x
-      report_link: "", //x
+      shortened_link: "",
+      report_link: "",
       date_created: new Date().toISOString().split("T")[0], //x
-      admin_testing_link: "", //x
+      admin_testing_link: "",
     },
   };
 
@@ -94,6 +94,73 @@ async function postFormDataToBackend(formData: RowType) {
   }
 }
 
+async function UpdateFormDataToBackend(
+  formData: PatchTypes,
+  sessionId: string
+) {
+  const { student, test, timeline, session } = formData;
+
+  let start_time = await formatDateTime(
+    timeline.startDate as string,
+    timeline.startTime as string
+  );
+  let end_time = await formatDateTime(
+    timeline.endDate as string,
+    timeline.endTime as string
+  );
+
+  let patchBody: any = {
+    start_time,
+    end_time,
+    signup_form_name: "Haryana Registration Form",
+    type: "sign-in",
+    auth_type: session.auth_type,
+    activate_signup: session.activate_signup,
+    id_generation: session.id_generation,
+    redirection: session.redirection,
+    pop_up_form: session.pop_up_form,
+    number_of_fields_in_pop_form: session.number_of_fields_in_pop_form,
+  };
+
+  if (student || test || timeline) {
+    patchBody.meta_data = {
+      ...(student.program ? { group: student.program } : {}),
+      ...(student.batch ? { batch: student.batch } : {}),
+      ...(student.grade ? { grade: student.grade } : {}),
+      ...(student.course ? { course: student.course } : {}),
+      ...(student.stream ? { stream: student.stream } : {}),
+      ...(test.format ? { test_format: test.format } : {}),
+      ...(test.purpose ? { test_purpose: test.purpose } : {}),
+      ...(test.cmsId ? { cms_test_id: test.cmsId } : {}),
+      ...(student.testTakers ? { test_takers_count: student.testTakers } : {}),
+      ...(test.optionalLimit ? { optional_limits: test.optionalLimit } : {}),
+      ...(test.markingScheme ? { marking_scheme: test.markingScheme } : {}),
+      ...(test.type ? { test_type: test.type } : {}),
+      ...(test.shortened_link ? { test_type: test.shortened_link } : {}),
+      date_created: timeline.date_created,
+      infinite_session: true,
+      has_synced_to_bq: false,
+    };
+  }
+
+  try {
+    const message = {
+      action: "patch",
+      id: sessionId,
+      patch_session: patchBody,
+    };
+
+    publishMessage(JSON.stringify(message));
+
+    return {
+      id: sessionId,
+    };
+  } catch (error) {
+    console.error("Error posting form data", error);
+    throw error;
+  }
+}
+
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
@@ -101,6 +168,18 @@ export default async function handler(
   if (req.method === "POST") {
     try {
       const response = await postFormDataToBackend(req.body);
+      res.status(200).json(response);
+    } catch (error) {
+      console.error("Error in API route", error);
+      res.status(500).json({ error: "Internal Server Error" });
+    }
+  } else if (req.method === "PATCH") {
+    try {
+      const sessionId = req.query.id;
+      const response = await UpdateFormDataToBackend(
+        req.body,
+        sessionId as string
+      );
       res.status(200).json(response);
     } catch (error) {
       console.error("Error in API route", error);
