+++ src/utils/FormInputHandling.ts
@@ -86,17 +86,17 @@ async function getASession(id: number) {
       platform: data.platform,
       markingScheme: data.meta_data?.marking_scheme,
       optionalLimit: data.meta_data?.optional_limits,
-      link: data.meta_data?.admin_testing_link,
+      link: data.meta_data?.admin_testing_link || null,
       id: data.id,
       sessionId: data.session_id,
-      sessionLink: data.meta_data?.shortened_link,
+      sessionLink: data.meta_data?.shortened_link || null,
       cmsId: data.meta_data?.cms_test_id,
     },
     timeline: {
-      isEnabled: data.meta_data?.enabled,
+      isEnabled: data.meta_data?.enabled || null,
       synced: data.meta_data?.has_synced_to_bq,
       sessionType: data.meta_data?.infinite_session,
-      reportLink: data.meta_data?.report_link,
+      reportLink: data.meta_data?.report_link || null,
       repeatSchedule: data.repeat_schedule,
       startDate: startDate,
       endDate: endDate,
