+++ tsconfig.json
@@ -18,6 +18,6 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "src/utils/PublishSnsMessage.ts", "src/pages/api/PostFormData.ts"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "src/utils/PublishSnsMessage.js", "src/pages/api/PostFormData.ts"],
   "exclude": ["node_modules"]
 }
